Refactor import preview helpers
